Added refresh token tests
