+++ src/CaptainHook.Common/Authentication/OAuthAuthenticationConfig.cs
@@ -30,6 +30,6 @@
         /// <summary>
         /// Refresh interval before the token expires
         /// </summary>
-        public int RefreshBeforeInSeconds { get; } = 10;
+        public int RefreshBeforeInSeconds { get; set; } = 10;
     }
 }
