--- src/CaptainHook.EventHandlerActor/Handlers/Authentication/IRefreshTokenHandler.cs
@@ -1,10 +1,0 @@
-﻿using System.Net.Http;
-using System.Threading.Tasks;
-
-namespace CaptainHook.EventHandlerActor.Handlers.Authentication
-{
-    public interface IRefreshTokenHandler
-    {
-        Task RefreshToken(HttpClient client);
-    }
-}
